Se actualizan vistas

+++ database/migrations/2016_12_06_000000_create_users_table.php
@@ -13,31 +13,33 @@ class CreateUsersTable extends Migration
      */
     public function up()
     {
-        Schema::create('users', function (Blueprint $table) {
-            $table->increments('id');
-            $table->string('firstname');
-            $table->string('lastname');
-            $table->string('email')->unique();
-            $table->string('phone')->nullable();
-            $table->string('password');
-            $table->string('imagesource')->nullable();
-            $table->string('businessName')->nullable();
-            $table->string('industry')->nullable();
-            $table->text('personaldescription')->nullable();
-            $table->string('city')->nullable();
-            $table->string('urlfacebook')->nullable();
-            $table->string('urllinkedin')->nullable();
-            $table->text('databank')->nullable();
-            $table->integer('tipo_clientes_id')->unsigned()->default(1);
-            $table->foreign('tipo_clientes_id')
-                  ->references('id')->on('tipo_clientes')
-                  ->onDelete('cascade');
-            $table->boolean('isAdmin')->default(false);
-            $table->rememberToken();
-            $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
-            $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
-            $table->softDeletes();
-        });
+        if (!Schema::hasTable('users')) {
+          Schema::create('users', function (Blueprint $table) {
+              $table->increments('id');
+              $table->string('firstname');
+              $table->string('lastname');
+              $table->string('email')->unique();
+              $table->string('phone')->nullable();
+              $table->string('password');
+              $table->string('imagesource')->nullable();
+              $table->string('businessName')->nullable();
+              $table->string('industry')->nullable();
+              $table->text('personaldescription')->nullable();
+              $table->string('city')->nullable();
+              $table->string('urlfacebook')->nullable();
+              $table->string('urllinkedin')->nullable();
+              $table->text('databank')->nullable();
+              $table->integer('tipo_clientes_id')->unsigned()->default(1);
+              $table->foreign('tipo_clientes_id')
+                    ->references('id')->on('tipo_clientes')
+                    ->onDelete('cascade');
+              $table->boolean('isAdmin')->default(false);
+              $table->rememberToken();
+              $table->timestamp('created_at')->default(\DB::raw('CURRENT_TIMESTAMP'));
+              $table->timestamp('updated_at')->default(\DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'));
+              $table->softDeletes();
+          });
+        }
     }
 
     /**
