+++ database/migrations/2016_12_12_023147_create_espacio_categoria_pivot_table.php
@@ -12,13 +12,15 @@ class CreateEspacioCategoriaPivotTable extends Migration
      */
     public function up()
     {
-        Schema::create('espacio_categoria', function (Blueprint $table) {
-            $table->integer('categoria_id')->unsigned()->index();
-            $table->foreign('categoria_id')->references('id')->on('categorias')->onDelete('cascade');
-            $table->integer('espacio_id')->unsigned()->index();
-            $table->foreign('espacio_id')->references('id')->on('espacios')->onDelete('cascade');
-            $table->primary(['categoria_id', 'espacio_id']);
-        });
+        if (!Schema::hasTable('categoria_espacio')) {
+            Schema::create('categoria_espacio', function (Blueprint $table) {
+                $table->integer('categoria_id')->unsigned()->index();
+                $table->foreign('categoria_id')->references('id')->on('categorias')->onDelete('cascade');
+                $table->integer('espacio_id')->unsigned()->index();
+                $table->foreign('espacio_id')->references('id')->on('espacios')->onDelete('cascade');
+                $table->primary(['categoria_id', 'espacio_id']);
+            });
+        }
     }
 
     /**
@@ -28,6 +30,6 @@ class CreateEspacioCategoriaPivotTable extends Migration
      */
     public function down()
     {
-        Schema::drop('categoria_espacio');
+        Schema::dropIfExists('categoria_espacio');
     }
 }
