+++ database/migrations/2016_12_20_193300_create_espacio_servicio_pivot_table.php
@@ -1,3 +1,4 @@
+
 <?php
 
 use Illuminate\Database\Schema\Blueprint;
@@ -12,13 +13,15 @@ class CreateEspacioServicioPivotTable extends Migration
      */
     public function up()
     {
-        Schema::create('espacio_servicio', function (Blueprint $table) {
-            $table->integer('espacio_id')->unsigned()->index();
-            $table->foreign('espacio_id')->references('id')->on('espacios')->onDelete('cascade');
-            $table->integer('servicio_id')->unsigned()->index();
-            $table->foreign('servicio_id')->references('id')->on('servicios')->onDelete('cascade');
-            $table->primary(['espacio_id', 'servicio_id']);
-        });
+        if (!Schema::hasTable('espacio_servicio')) {
+            Schema::create('espacio_servicio', function (Blueprint $table) {
+                $table->integer('espacio_id')->unsigned()->index();
+                $table->foreign('espacio_id')->references('id')->on('espacios')->onDelete('cascade');
+                $table->integer('servicio_id')->unsigned()->index();
+                $table->foreign('servicio_id')->references('id')->on('servicios')->onDelete('cascade');
+                $table->primary(['espacio_id', 'servicio_id']);
+            });
+        }
     }
 
     /**
@@ -28,6 +31,6 @@ class CreateEspacioServicioPivotTable extends Migration
      */
     public function down()
     {
-        Schema::drop('espacio_servicio');
+        Schema::dropIfExists('espacio_servicio');
     }
 }
