+++ database/migrations/2016_12_20_193527_create_espacio_tag_pivot_table.php
@@ -12,13 +12,15 @@ class CreateEspacioTagPivotTable extends Migration
      */
     public function up()
     {
-        Schema::create('espacio_tag', function (Blueprint $table) {
-            $table->integer('espacio_id')->unsigned()->index();
-            $table->foreign('espacio_id')->references('id')->on('espacios')->onDelete('cascade');
-            $table->integer('tag_id')->unsigned()->index();
-            $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
-            $table->primary(['espacio_id', 'tag_id']);
-        });
+        if (!Schema::hasTable('espacio_tag')) {
+            Schema::create('espacio_tag', function (Blueprint $table) {
+                $table->integer('espacio_id')->unsigned()->index();
+                $table->foreign('espacio_id')->references('id')->on('espacios')->onDelete('cascade');
+                $table->integer('tag_id')->unsigned()->index();
+                $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade');
+                $table->primary(['espacio_id', 'tag_id']);
+            });
+        }
     }
 
     /**
@@ -28,6 +30,6 @@ class CreateEspacioTagPivotTable extends Migration
      */
     public function down()
     {
-        Schema::drop('espacio_tag');
+        Schema::dropIfExists('espacio_tag');
     }
 }
